refactor: split responses input normalizer helpers
